#151 moved commands from view to interactor
