+++ shvatka/infrastructure/di/interactors.py
@@ -26,7 +26,7 @@ from shvatka.core.services.current_game import CurrentGameProviderImpl
 from shvatka.core.services.key import KeyProcessor, TimerProcessor
 from shvatka.core.services.game_play import CheckKeyInteractor
 from shvatka.core.waiver.interactors import WaiversReaderInteractor, AddWaiverVoteInteractor
-from shvatka.infrastructure.db.dao.complex import WaiverVoteAdderImpl, WaiverVoteGetterImpl
+from shvatka.infrastructure.db.dao.complex2.waiver import WaiverVoteAdderImpl, WaiverVoteGetterImpl
 from shvatka.infrastructure.db.dao.complex.game import GameFilesGetterImpl, GamePlayReaderImpl
 from shvatka.infrastructure.db.dao.complex.game import (
     GameFilesGetterImpl,
